+++ bdb/src/main/java/org/judal/bdb/DBTableDataSource.java
@@ -8,6 +8,11 @@ import java.util.Properties;
 
 import javax.jdo.JDOException;
 import javax.jdo.JDOUnsupportedOptionException;
+import javax.transaction.HeuristicMixedException;
+import javax.transaction.HeuristicRollbackException;
+import javax.transaction.NotSupportedException;
+import javax.transaction.RollbackException;
+import javax.transaction.SystemException;
 import javax.transaction.TransactionManager;
 
 import org.judal.metadata.ColumnDef;
@@ -67,7 +72,7 @@ public class DBTableDataSource extends DBDataSource implements TableDataSource {
 		DBTable retval;
 		Properties oProps = getTableProperties(recordInstance.getTableName(), new ColumnDef[0]);
 		oProps.put("readonly", isReadOnly() ? "true" : "false");
-		retval = (DBTable) openTableOrBucket(oProps, getMetaData().getTable(recordInstance.getTableName()), recordInstance.getClass());
+		retval = (DBTable) openTableOrBucket(oProps, getMetaData().getTable(recordInstance.getTableName()), recordInstance.getClass(), true);
 
 		if (DebugFile.trace) {
 			DebugFile.decIdent();
@@ -89,7 +94,7 @@ public class DBTableDataSource extends DBDataSource implements TableDataSource {
 		DBTable retval;
 		Properties oProps = getTableProperties(recordInstance.getTableName(), new ColumnDef[0]);
 		oProps.put("readonly", "true");
-		retval = (DBTable) openTableOrBucket(oProps, getMetaData().getTable(recordInstance.getTableName()), recordInstance.getClass());
+		retval = (DBTable) openTableOrBucket(oProps, getMetaData().getTable(recordInstance.getTableName()), recordInstance.getClass(), true);
 
 		if (DebugFile.trace) {
 			DebugFile.decIdent();
@@ -111,7 +116,7 @@ public class DBTableDataSource extends DBDataSource implements TableDataSource {
 		DBTable retval;
 		Properties oProps = getTableProperties(recordInstance.getTableName(), recordInstance.columns());
 		oProps.put("readonly", isReadOnly() ? "true" : "false");
-		retval = (DBTable) openTableOrBucket(oProps, getMetaData().getTable(recordInstance.getTableName()), recordInstance.getClass());
+		retval = (DBTable) openTableOrBucket(oProps, getMetaData().getTable(recordInstance.getTableName()), recordInstance.getClass(), true);
 
 		if (DebugFile.trace) {
 			DebugFile.decIdent();
@@ -133,7 +138,7 @@ public class DBTableDataSource extends DBDataSource implements TableDataSource {
 		DBTable retval;
 		Properties oProps = getTableProperties(recordInstance.getTableName(), recordInstance.columns());
 		oProps.put("readonly", "true");
-		retval =  (DBTable) openTableOrBucket(getTableProperties(recordInstance.getTableName(), recordInstance.columns()), getMetaData().getTable(recordInstance.getTableName()), recordInstance.getClass());
+		retval =  (DBTable) openTableOrBucket(getTableProperties(recordInstance.getTableName(), recordInstance.columns()), getMetaData().getTable(recordInstance.getTableName()), recordInstance.getClass(), true);
 
 		if (DebugFile.trace) {
 			DebugFile.decIdent();
@@ -147,8 +152,18 @@ public class DBTableDataSource extends DBDataSource implements TableDataSource {
 
 	@Override
 	public void createTable(TableDef tableDef, Map<String,Object> options) throws JDOException {
-		DBTable oTbl = (DBTable) openTableOrBucket(getTableProperties(tableDef.getName(), tableDef.getColumns()), getMetaData().getTable(tableDef.getName()), MapRecord.class);
-	  	oTbl.close();
+		if (inTransaction())
+			throw new JDOException("Cannot create a table in the middle of a transaction");
+		try {
+			if (isTransactional())
+				getTransactionManager().begin();
+			DBTable oTbl = (DBTable) openTableOrBucket(getTableProperties(tableDef.getName(), tableDef.getColumns()), tableDef, tableDef.getRecordClass(), isTransactional());
+			if (isTransactional())
+				getTransactionManager().commit();
+			oTbl.close();
+		} catch (NotSupportedException | SystemException | SecurityException | IllegalStateException | RollbackException | HeuristicMixedException | HeuristicRollbackException | IllegalArgumentException | ClassNotFoundException e) {
+			throw new JDOException(e.getMessage(), e);
+		}
 	}
 
 	// --------------------------------------------------------------------------
@@ -185,6 +200,8 @@ public class DBTableDataSource extends DBDataSource implements TableDataSource {
 	
 	@Override
 	public void dropTable(String sDbk, boolean cascade) throws JDOException {
+		if (inTransaction())
+			throw new JDOException("Cannot drop a table in the middle of a transaction");
 		if (DebugFile.trace) {
 			DebugFile.writeln("Begin DBTableDataSource.dropTable("+sDbk+", "+String.valueOf(cascade)+")");
 			DebugFile.incIdent();
@@ -192,16 +209,26 @@ public class DBTableDataSource extends DBDataSource implements TableDataSource {
 		try {
 			Properties oProps = getTableProperties(sDbk, getMetaData().getTable(sDbk).getColumns());
 			oProps.put("readonly","false");
-			DBTable oTbl = (DBTable) openTableOrBucket(oProps, getMetaData().getTable(sDbk), Record.class);
+			DBTable oTbl = null;
+			if (isTransactional())
+				getTransactionManager().begin();
+			oTbl = (DBTable) openTableOrBucket(oProps, getMetaData().getTable(sDbk), getMetaData().getTable(sDbk).getRecordClass(), isTransactional());
 			for (DBIndex oInd : oTbl.indexes())
 				oTbl.dropIndex(oInd.getName());
 			oTbl.close();
 			if (DebugFile.trace)
 				DebugFile.writeln("removeDatabase("+getPath()+sDbk+".db, "+sDbk+")");
 			getEnvironment().removeDatabase(getTransaction(), getPath()+sDbk+".db", sDbk);
+			if (DebugFile.trace)
+				DebugFile.writeln("database "+getPath()+sDbk+".db, removed");
+			if (isTransactional())
+				getTransactionManager().commit();
 			File oDbf = new File(getPath()+sDbk+".db");
-			if (oDbf.exists()) oDbf.delete();
-		} catch (FileNotFoundException | DatabaseException e) {
+			if (oDbf.exists()) {
+				if (DebugFile.trace) DebugFile.writeln("deleting file "+getPath()+sDbk+".db");
+				oDbf.delete();
+			}
+		} catch (FileNotFoundException | DatabaseException | IllegalArgumentException | IllegalStateException | ClassNotFoundException | NotSupportedException | SystemException | SecurityException | RollbackException | HeuristicMixedException | HeuristicRollbackException e) {
 			throw new JDOException(e.getMessage(), e);
 		}
 		if (DebugFile.trace) {
@@ -214,9 +241,25 @@ public class DBTableDataSource extends DBDataSource implements TableDataSource {
 	
 	@Override
 	public void truncateTable(String tableName, boolean cascade) throws JDOException {
-		DBTable tbl = (DBTable) openTableOrBucket(getTableProperties(tableName, getMetaData().getColumns(tableName)), getMetaData().getTable(tableName), MapRecord.class);
+		if (inTransaction())
+			throw new JDOException("Cannot truncate a table in the middle of a transaction");
+
+		if (isTransactional())
+			try {
+				getTransactionManager().begin();
+			} catch (NotSupportedException | SystemException e) {
+				throw new JDOException(e.getMessage(), e);
+			}
+
+		TableDef tdef = getMetaData().getTable(tableName);
+		DBTable tbl = null;
 		try {
-			tbl.truncate();
+			tbl = (DBTable) openTableOrBucket(getTableProperties(tableName, getMetaData().getColumns(tableName)), tdef, tdef.getRecordClass(), isTransactional());
+			tbl.truncate(isTransactional());
+			if (isTransactional())
+				getTransactionManager().commit();
+		} catch (SecurityException | IllegalStateException | RollbackException | HeuristicMixedException | HeuristicRollbackException | SystemException | ArrayIndexOutOfBoundsException | IllegalArgumentException | ClassNotFoundException e) {
+			throw new JDOException(e.getMessage(), e);
 		} finally {
 			if (tbl!=null) tbl.close();
 		}
